Move token amount utils to separate module
