+++ plt/plt-scheduler/src/protocol_level_tokens/token_module/initialize.rs
@@ -2,17 +2,15 @@ use crate::block_state_polymorph::token::TokenPXRefMut;
 use crate::failure::{
     HigherLevelProtocolError, ResultWithBlockStateFailure, ResultWithBlockStateFailureExt,
 };
-use crate::protocol_level_tokens::balance_operations;
-use crate::protocol_level_tokens::token_module::errors::{
-    MintWouldOverflowError, TokenAmountDecimalsMismatchError,
-};
-use crate::protocol_level_tokens::token_module::util;
+use crate::protocol_level_tokens::{balance_operations, token_amount};
 use concordium_base::common::cbor::CborSerializationError;
 use concordium_base::protocol_level_tokens::{TokenAdminRole, TokenModuleInitializationParameters};
 use plt_block_state::entity::accounts::Accounts;
 use plt_block_state::entity::{EntityContext, EntityContextTypes};
 use plt_block_state::external::AccountNotFoundByAddressError;
 use plt_scheduler_types::types::events::BlockItemEvent;
+use crate::protocol_level_tokens::balance_operations::MintWouldOverflowError;
+use crate::protocol_level_tokens::token_amount::TokenAmountDecimalsMismatchError;
 
 /// Represents the reasons why [`initialize_token`] can fail.
 #[derive(Debug, thiserror::Error)]
@@ -96,7 +94,7 @@ pub fn initialize_token<C: EntityContextTypes>(
         .set_governance_account(context, governance_account_index)?;
 
     if let Some(initial_supply) = init_params.initial_supply {
-        let mint_amount = util::to_raw_token_amount(&token_configuration, initial_supply)?;
+        let mint_amount = token_amount::to_raw_token_amount(&token_configuration, initial_supply)?;
 
         balance_operations::mint(
             context,
